Move mobile sidenav file, pass groupId params
