+++ src/components/Sidenav/SidenavGroupUI.vue
@@ -17,7 +17,7 @@
         flat
         dense
         round
-        :to="{name: 'groupSettings'}"
+        :to="{ name: 'groupSettings', params: { groupId } }"
       >
         <q-icon name="fas fa-cog fa-fw" />
         <q-tooltip v-t="'GROUP.SETTINGS'" />
@@ -38,7 +38,7 @@
         no-border
         class="no-padding"
       >
-        <q-item :to="{name: 'group'}">
+        <q-item :to="{ name: 'group', params: { groupId } }">
           <q-item-side class="text-center">
             <q-icon name="fas fa-bullhorn" />
           </q-item-side>
@@ -46,7 +46,7 @@
             {{ $t("GROUP.WALL") }}
           </q-item-main>
         </q-item>
-        <q-item :to="{name: 'groupPickups'}">
+        <q-item :to="{ name: 'groupPickups', params: { groupId } }">
           <q-item-side class="text-center">
             <q-icon name="fas fa-shopping-basket" />
           </q-item-side>
@@ -54,7 +54,7 @@
             {{ $t("GROUP.PICKUPS") }}
           </q-item-main>
         </q-item>
-        <q-item :to="{name: 'groupFeedback'}">
+        <q-item :to="{ name: 'groupFeedback', params: { groupId } }">
           <q-item-side class="text-center">
             <q-icon name="fas fa-balance-scale" />
           </q-item-side>
@@ -62,7 +62,7 @@
             {{ $t("PICKUP_FEEDBACK.TITLE") }}
           </q-item-main>
         </q-item>
-        <q-item :to="{name: 'groupDescription'}">
+        <q-item :to="{ name: 'groupDescription', params: { groupId } }">
           <q-item-side class="text-center">
             <i class="far fa-address-card"/>
           </q-item-side>
@@ -70,7 +70,7 @@
             {{ $t("GROUP.DESCRIPTION") }}
           </q-item-main>
         </q-item>
-        <q-item :to="{name: 'groupMembers'}">
+        <q-item :to="{ name: 'groupMembers', params: { groupId } }">
           <q-item-side class="text-center">
             <q-icon name="fas fa-users" />
           </q-item-side>
@@ -78,7 +78,7 @@
             {{ $t("GROUP.MEMBERS") }}
           </q-item-main>
         </q-item>
-        <q-item :to="{name: 'groupHistory'}">
+        <q-item :to="{ name: 'groupHistory', params: { groupId } }">
           <q-item-side class="text-center">
             <i class="far fa-clock"/>
           </q-item-side>
@@ -88,7 +88,7 @@
         </q-item>
         <q-item
           v-if="$q.platform.is.mobile"
-          :to="{name: 'map'}"
+          :to="{ name: 'map', params: { groupId } }"
         >
           <q-item-side class="text-center">
             <q-icon name="fas fa-map" />
@@ -112,6 +112,7 @@ export default {
     SidenavBox, GroupOptions, QBtn, QList, QItem, QItemSide, QItemMain, QIcon, QTooltip,
   },
   props: {
+    groupId: { required: true, type: Number },
     expanded: { default: true, type: Boolean },
   },
 }
