+++ src/pages/MainLayout.vue
@@ -103,7 +103,7 @@
 import KTopbar from '@/components/Layout/KTopbar'
 import KTopbarLoggedOut from '@/components/Layout/LoggedOut/KTopbar'
 import KFooter from '@/components/Layout/KFooter'
-import MobileSidenav from '@/components/Layout/MobileSidenav'
+import MobileSidenav from '@/components/Sidenav/MobileSidenav'
 import Banners from '@/components/Layout/Banners'
 import RouteError from '@/components/RouteError'
 import UnsupportedBrowserWarning from '@/components/UnsupportedBrowserWarning'
